+++ docs/markdown/ios/button.md
@@ -48,7 +48,7 @@ People use Apple’s accessibility features, such as reduced transparency, Voice
 
 ### Size
 
-Mobile buttons are available in 2 sizes. The Button text always use [$font-size-300 token](/foundations/design_tokens#Font-size) (16sp).
+Mobile buttons are available in 2 sizes. The Button text always use [$font-size-300 token](/foundations/design_tokens/overview#Font-size) (16sp).
 
 1. **lg (60px)**
    Large should be primarily used on Pinner, business and internal surfaces.
