+++ docs/pages/web/heading.js
@@ -274,7 +274,7 @@ Avoid overriding our line-height settings, as this can result in text clipping f
 
       <MainSection name="Variants">
         <MainSection.Subsection
-          description={`These font sizes follow those available through our [Design Tokens](/foundations/design_tokens#Font-size). If your text does not need to be a [semantic heading (H1-H6)](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements), use [Text](/web/text) instead.
+          description={`These font sizes follow those available through our [Design Tokens](/foundations/design_tokens/overview#Font-size). If your text does not need to be a [semantic heading (H1-H6)](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements), use [Text](/web/text) instead.
 `}
           title="Size"
         >
