+++ docs/redirects.js
@@ -163,6 +163,11 @@ const misc = [
     permanent: true,
   },
   {
+    source: '/foundations/design_tokens',
+    destination: '/foundations/design_tokens/overview',
+    permanent: true,
+  },
+  {
     source: '/foundations/rtl_guidelines/about_international_design',
     destination: '/foundations/international_design/about_international_design',
     permanent: true,
