+++ playwright/accessibility/foundations__design_tokens__component_tokens.spec.mjs
@@ -3,6 +3,6 @@ import { test } from '@playwright/test';
 import expectAccessiblePage from './expectAccessiblePage.mjs';
 
 test('Design Tokens Accessibility check', async ({ page }) => {
-  await page.goto('/foundations/design_tokens');
+  await page.goto('/foundations/design_tokens/overview');
   await expectAccessiblePage({ page });
 });
