+++ playwright/accessibility/foundations__design_tokens__overview.spec copy.mjs
@@ -1,0 +1,8 @@
+// @flow strict
+import { test } from '@playwright/test';
+import expectAccessiblePage from './expectAccessiblePage.mjs';
+
+test('Design Component Tokens Accessibility check', async ({ page }) => {
+  await page.goto('/foundations/design_tokens/component_tokens');
+  await expectAccessiblePage({ page });
+});
